+++ examples/teensy4/src/configured.rs
@@ -31,6 +31,7 @@ fn main() -> ! {
     let bus = usb_device::bus::UsbBusAllocator::new(bus_adapter);
     let mut device = UsbDeviceBuilder::new(&bus, UsbVidPid(0x5824, 0x27dd))
         .product("imxrt-usbd")
+        .max_packet_size_0(64)
         .build();
 
     loop {