+++ src/bus.rs
@@ -9,7 +9,7 @@
 //!
 //! Most of the interesting behavior happens in the driver.
 
-use super::driver::FullSpeed;
+use super::driver::Driver;
 use crate::gpt;
 use core::cell::RefCell;
 use cortex_m::interrupt::{self, Mutex};
@@ -19,9 +19,11 @@ use usb_device::{
     UsbDirection,
 };
 
-/// A `UsbBus` implementation
+pub use super::driver::Speed;
+
+/// A full- and high-speed `UsbBus` implementation
 ///
-/// The `BusAdapter` adapts the RAL instances, and exposes a `UsbBus` implementation.
+/// The `BusAdapter` adapts the USB peripheral instances, and exposes a `UsbBus` implementation.
 ///
 /// # Requirements
 ///
@@ -42,7 +44,7 @@ use usb_device::{
 /// your USB class' documentation for details. This example also skips the clock initialization.
 ///
 /// ```no_run
-/// use imxrt_usbd::full_speed::BusAdapter;
+/// use imxrt_usbd::BusAdapter;
 ///
 /// # struct Ps; use imxrt_usbd::Instance as Inst;
 /// # unsafe impl imxrt_usbd::Peripherals for Ps { fn instance(&self) -> Inst { panic!() } }
@@ -80,12 +82,46 @@ use usb_device::{
 ///
 /// // Ready for class traffic!
 /// ```
+///
+/// # Design
+///
+/// This section talks about the driver design. It assumes that
+/// you're familiar with the details of the i.MX RT USB peripheral. If you
+/// just want to use the driver, you can skip this section.
+///
+/// ## Packets and transfers
+///
+/// All i.MX RT USB drivers manage queue heads (QH), and transfer
+/// descriptors (TD). For the driver, each (QH) is assigned
+/// only one (TD) to perform I/O. We then assume each TD describes a single
+/// packet. This is simple to implement, but it means that the
+/// driver can only have one packet in flight per endpoint. You're expected
+/// to quickly respond to `poll()` outputs, and schedule the next transfer
+/// in the time required for devices. This becomes more important as you
+/// increase driver speeds.
+///
+/// The hardware can zero-length terminate (ZLT) packets as needed if you
+/// call [`enable_zlt`](BusAdapter::enable_zlt). By default, this feature is
+/// off, because most `usb-device` classes / devices take care to send zero-length
+/// packets, and enabling this feature could interfere with the class / device
+/// behaviors.
 pub struct BusAdapter {
-    usb: Mutex<RefCell<FullSpeed>>,
+    usb: Mutex<RefCell<Driver>>,
+    cs: Option<cortex_m::interrupt::CriticalSection>,
 }
 
 impl BusAdapter {
-    /// Create a USB bus adapter
+    /// Create a high-speed USB bus adapter
+    ///
+    /// This is equivalent to [`BusAdapter::with_speed`] when supplying [`Speed::High`]. See
+    /// the `with_speed` documentation for more information.
+    pub fn new<P: crate::Peripherals>(peripherals: P, buffer: &'static mut [u8]) -> Self {
+        Self::with_speed(peripherals, buffer, Speed::High)
+    }
+
+    /// Create a USB bus adapter with the given speed
+    ///
+    /// Specify [`Speed::LowFull`] to throttle the USB data rate.
     ///
     /// When this function returns, the `BusAdapter` has initialized the PHY and USB core peripherals.
     /// The adapter expects to own these two peripherals, along with the other peripherals required
@@ -95,17 +131,53 @@ impl BusAdapter {
     /// memory region will be partitioned for the endpoints, based on their requirements.
     ///
     /// You must ensure that no one else is using the endpoint memory!
-    pub fn new<P: crate::Peripherals>(peripherals: P, buffer: &'static mut [u8]) -> Self {
-        let mut usb = FullSpeed::new(peripherals);
+    pub fn with_speed<P: crate::Peripherals>(
+        peripherals: P,
+        buffer: &'static mut [u8],
+        speed: Speed,
+    ) -> Self {
+        Self::init(peripherals, buffer, speed, None)
+    }
 
-        usb.initialize();
+    /// Create a USB bus adapter that never takes a critical section
+    ///
+    /// See [`BusAdapter::with_speed`] for general information.
+    ///
+    /// # Safety
+    ///
+    /// The returned object fakes its `Sync` safety. Specifically, the object
+    /// will not take critical sections in its `&[mut] self` methods to ensure safe
+    /// access. By using this object, you must manually hold the guarantees of
+    /// `Sync` without the compiler's help.
+    pub unsafe fn without_critical_sections<P: crate::Peripherals>(
+        peripherals: P,
+        buffer: &'static mut [u8],
+        speed: Speed,
+    ) -> Self {
+        Self::init(
+            peripherals,
+            buffer,
+            speed,
+            Some(cortex_m::interrupt::CriticalSection::new()),
+        )
+    }
+
+    fn init<P: crate::Peripherals>(
+        peripherals: P,
+        buffer: &'static mut [u8],
+        speed: Speed,
+        cs: Option<cortex_m::interrupt::CriticalSection>,
+    ) -> Self {
+        let mut usb = Driver::new(peripherals);
+
+        usb.initialize(speed);
         usb.set_endpoint_memory(buffer);
 
         BusAdapter {
             usb: Mutex::new(RefCell::new(usb)),
+            cs,
         }
     }
-
     /// Enable (`true`) or disable (`false`) interrupts for this USB peripheral
     ///
     /// The interrupt causes are implementation specific. To handle the interrupt,
@@ -114,22 +186,48 @@ impl BusAdapter {
         self.with_usb_mut(|usb| usb.set_interrupts(interrupts));
     }
 
-    /// Interrupt-safe, immutable access to the USB peripheral
-    fn with_usb<R>(&self, func: impl FnOnce(&FullSpeed) -> R) -> R {
-        interrupt::free(|cs| {
+    /// Enable zero-length termination (ZLT) for the given endpoint
+    ///
+    /// When ZLT is enabled, software does not need to send a zero-length packet
+    /// to terminate a transfer where the number of bytes equals the max packet size.
+    /// The hardware will send this zero-length packet itself. By default, ZLT is off,
+    /// and software is expected to send these packets. Enable this if you're confident
+    /// that your (third-party) device / USB class isn't already sending these packets.
+    ///
+    /// # Panics
+    ///
+    /// `enable_zlt` must be called before the endpoint is allocated. If the endpoint is
+    /// already allocated, this call panics.
+    pub fn enable_zlt(&self, ep_addr: EndpointAddress) {
+        self.with_usb_mut(|usb| usb.enable_zlt(ep_addr));
+    }
+
+    /// Immutable access to the USB peripheral
+    fn with_usb<R>(&self, func: impl FnOnce(&Driver) -> R) -> R {
+        let with_cs = |cs: &'_ _| {
             let usb = self.usb.borrow(cs);
             let usb = usb.borrow();
             func(&usb)
-        })
+        };
+        if let Some(cs) = &self.cs {
+            with_cs(cs)
+        } else {
+            interrupt::free(with_cs)
+        }
     }
 
-    /// Interrupt-safe, mutable access to the USB peripheral
-    fn with_usb_mut<R>(&self, func: impl FnOnce(&mut FullSpeed) -> R) -> R {
-        interrupt::free(|cs| {
+    /// Mutable access to the USB peripheral
+    fn with_usb_mut<R>(&self, func: impl FnOnce(&mut Driver) -> R) -> R {
+        let with_cs = |cs: &'_ _| {
             let usb = self.usb.borrow(cs);
             let mut usb = usb.borrow_mut();
             func(&mut usb)
-        })
+        };
+        if let Some(cs) = &self.cs {
+            with_cs(cs)
+        } else {
+            interrupt::free(with_cs)
+        }
     }
 
     /// Apply device configurations, and perform other post-configuration actions
@@ -163,6 +261,18 @@ impl BusAdapter {
     ) -> R {
         let usb = self.usb.borrow(cs);
         usb.borrow_mut().gpt_mut(instance, func)
+    }
+
+    /// Acquire one of the GPT timer instances.
+    ///
+    /// `instance` identifies which GPT instance you're accessing.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the GPT instance is already borrowed. This could happen
+    /// if you call `borrow_gpt` again within the `func` callback.
+    pub fn gpt_mut<R>(&self, instance: gpt::Instance, func: impl FnOnce(&mut gpt::Gpt) -> R) -> R {
+        self.with_usb_mut(|usb| usb.gpt_mut(instance, func))
     }
 }